% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataPreprocess.R
\name{changeColNames}
\alias{changeColNames}
\title{change column names}
\usage{
changeColNames(x, ind, newNames)
}
\arguments{
\item{x}{\itemize{
\item Data matrix
}}

\item{ind}{\itemize{
\item vector of index of colnames needed to be changed
}}

\item{newNames}{\itemize{
\item new column names
}}
}
\value{
data matrix with changed column names
}
\description{
change column names
}
