% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataPreprocess.R
\name{cleanUpRownames}
\alias{cleanUpRownames}
\title{Make legal row/column names}
\usage{
cleanUpRownames(mat, useRownames = FALSE)
}
\arguments{
\item{mat}{character vector of names to make legal}
}
\value{
The legal row/column names
}
\description{
make legal row/column names
}
\examples{
\dontrun{
  rown <- makeLegalRowname(mets)
  }

}
