% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataPreprocess.R
\name{combineDatasets}
\alias{combineDatasets}
\title{Combine datasets}
\usage{
combineDatasets(ref, research)
}
\arguments{
\item{ref}{\itemize{
\item Current data matrix
}}

\item{research}{\itemize{
\item Data matrix you want to combine with ref.
}}
}
\value{
combined.data - Combined data matrix.
}
\description{
Combine datasets
}
\examples{
# Row names and column names are required for both input matrices.
ref=matrix(rnorm(500), ncol=100)
rownames(ref)=sprintf("Feature\%d",sample(seq_len(20),
                                nrow(ref),replace = FALSE))
colnames(ref)=sprintf("Sample\%d", seq_len(ncol(ref)))
research=matrix(rnorm(500), ncol=100)
rownames(research)=sprintf("Feature\%d",sample(seq_len(20),
                                nrow(ref),replace = FALSE))
colnames(research) = sprintf("Sample\%d", seq_len(ncol(ref)))
combined.data = data.combineData(ref, research)
}
