% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GraphModel.R
\name{disFromDowntown}
\alias{disFromDowntown}
\title{get distance to disease module (or other node set)}
\usage{
disFromDowntown(dis_mod, ptBSbyK.dis, p2.sig.nodes, p2.optBS, ranks, G)
}
\arguments{
\item{dis_mod}{\itemize{
\item a vector of disease module
}}
}
\description{
get distance to disease module (or other node set)
}
