% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataPreprocess.R
\name{formatNumDataset}
\alias{formatNumDataset}
\title{Format and numerize dataset}
\usage{
formatNumDataset(x)
}
\arguments{
\item{x}{\itemize{
\item Data matrix
}}
}
\value{
zscored.data - Z-transformed data.
}
\description{
Format and numerize dataset
}
