% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatAnalysis.R
\name{getAssociation}
\alias{getAssociation}
\title{get association}
\usage{
getAssociation(D, ress, preds, ctrlVs = NULL, padj.method = "bonferroni")
}
\arguments{
\item{D}{The data frame containing predictors, responses and covariates}

\item{ress}{vector containing all names of all response}

\item{preds}{vector containing all names of all predictors}

\item{ctrlVs}{varaibles that need to be controled for}

\item{padjMethod}{p-value adjustment method}
}
\value{
A data frame of coeffecients and p-values
}
\description{
perform profile-wise association analysis
}
\examples{
assc = getAssociatio(D, ress=c("age","gender"), preds=grep("feature",colnames(D),value=T));
}
