% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GraphModel.R
\name{getCTDmodule}
\alias{getCTDmodule}
\title{getCTDmodule}
\usage{
getCTDmodule(
  ig,
  profile,
  kmx = 30,
  useRanks = T,
  ranks = NULL,
  useS = F,
  S = NULL,
  p1 = 0.9,
  thresholdDiff = 0.01
)
}
\arguments{
\item{data_mx}{\itemize{
\item Normalized, imputed, z-scored data. Data matrix includes features as rows, samples as columns.
}}

\item{cases}{\itemize{
\item vector of index or column names corresponding to cases
}}

\item{controls}{\itemize{
\item vector of index or column names corresponding to controls
}}

\item{fillRateThreashold}{\itemize{
\item percentage of allowed minimum non-NA value percentage per feature.
}}

\item{useIgRef}{\itemize{
\item Boolean. If TRUE, provide ig_ref as graph learned from references
}}

\item{ig_ref}{\itemize{
\item graph learned from references
}}
}
\value{
learned network in igraph object
}
\description{
learn condition-specific models
}
