% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatAnalysis.R
\name{getFill}
\alias{getFill}
\title{Get feature fill rate}
\usage{
getFill(x, byColumn = F)
}
\arguments{
\item{x}{The data matrix or data frame}

\item{byColumn}{Boolean: whether the function done by column}
}
\value{
The fill rate
}
\description{
get percentage of non NA cells per row or per column
}
\examples{
fill <- getFill(data_mx, byColumn=F);
}
