% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatAnalysis.R
\name{getPertubedByColumn}
\alias{getPertubedByColumn}
\title{Get rate of abnormal measures by column}
\usage{
getPertubedByColumn(x, threshold)
}
\arguments{
\item{x}{The data matrix or data frame}

\item{threshold}{threshold where measures be count as abnormal if its absolute value is above the or below the threshold}
}
\value{
The abnormal rate
}
\description{
get percentage of abnormal measures per column
}
\examples{
fill <- getPertubedByColumn(data_mx, threshold=2);
}
