% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataPreprocess.R
\name{imputeMissingValues}
\alias{imputeMissingValues}
\title{Impute missing values as minimum}
\usage{
imputeMissingValues(data, ref)
}
\arguments{
\item{data}{\itemize{
\item Normalized data matrix. Features as rows, samples as columns.
}}

\item{ref}{\itemize{
\item Normalized reference data matrix. Features as rows, samples as columns.
}}
}
\value{
imputed data
}
\description{
Impute using uniform random variable, where a = 0.99*observed minimum, and b = observed minimum
}
\examples{
zscored.data=imputeMissingValues(dis_data,ref_data)
}
