% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GraphModel.R
\name{learnModel}
\alias{learnModel}
\title{learn condition-specific models}
\usage{
learnModel(
  data_mx,
  cases,
  controls,
  fillRateThreashold,
  useIgRef = F,
  ig_ref = NULL
)
}
\arguments{
\item{data_mx}{\itemize{
\item Normalized, imputed, z-scored data. Data matrix includes features as rows, samples as columns.
}}

\item{cases}{\itemize{
\item vector of index or column names corresponding to cases
}}

\item{controls}{\itemize{
\item vector of index or column names corresponding to controls
}}

\item{fillRateThreashold}{\itemize{
\item percentage of allowed minimum non-NA value percentage per feature.
}}

\item{useIgRef}{\itemize{
\item Boolean. If TRUE, provide ig_ref as graph learned from references
}}

\item{ig_ref}{\itemize{
\item graph learned from references
}}
}
\value{
learned network in igraph object
}
\description{
learn condition-specific models
}
\examples{
data(Miller2015)
data_mx.og = as.matrix(Miller2015[,grep("IEM_", colnames(Miller2015))])# one sample per column, one metabolite per row.
diagnoses=data_mx.og["diagnosis",]
cohorts = list()
cohorts$msud = colnames(data_mx.og[,which(data_mx.og["diagnosis",]=="Maple syrup urine disease")])
cohorts$pku = colnames(data_mx.og[,which(data_mx.og["diagnosis",]=="Phenylketonuria")])
cohorts$ref = colnames(data_mx.og[,which(data_mx.og["diagnosis",]=="No biochemical genetic diagnosis")])
data_mx=apply(data_mx.og[unlist(cohorts)],c(1,2),as.numeric)
ig_ref=learnPartialCorrelationGraph(data_mx, cohorts$ref, cohorts$ref, fillRateThreashold=0.8)
ig_PKU=learnModel(data_mx, cohorts$pku, cohorts$ref, fillRateThreashold=0.8, useIgRef=T, ig_ref=ig_ref)
}
