% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GraphModel.R
\name{learnPartialCorrelationGraph}
\alias{learnPartialCorrelationGraph}
\title{learn partial correlation graph}
\usage{
learnPartialCorrelationGraph(data_mx, cases, controls, fillRateThreashold)
}
\arguments{
\item{data_mx}{\itemize{
\item Normalized, imputed, z-scored data. Data matrix includes features as rows, samples as columns.
}}

\item{cases}{\itemize{
\item vector of index or column names corresponding to cases
}}

\item{controls}{\itemize{
\item vector of index or column names corresponding to controls
}}

\item{fillRateThreashold}{\itemize{
\item percentage of allowed minimum non-NA value percentage per feature in cases and controls.
}}
}
\value{
learned network in igraph object
}
\description{
learn partial correlation-based graphic models
}
\examples{
data(Miller2015)
data_mx.og = as.matrix(Miller2015[,grep("IEM_", colnames(Miller2015))])# one sample per column, one metabolite per row.
diagnoses=data_mx.og["diagnosis",]
cohorts = list()
cohorts$msud = colnames(data_mx.og[,which(data_mx.og["diagnosis",]=="Maple syrup urine disease")])
cohorts$pku = colnames(data_mx.og[,which(data_mx.og["diagnosis",]=="Phenylketonuria")])
cohorts$ref = colnames(data_mx.og[,which(data_mx.og["diagnosis",]=="No biochemical genetic diagnosis")])
data_mx=apply(data_mx.og[unlist(cohorts)],c(1,2),as.numeric)
ig=learnPartialCorrelationGraph(data_mx, cohorts$pku, cohorts$ref, fillRateThreashold=0.8)
}
