% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataPreprocess.R
\name{makeLegalRowname}
\alias{makeLegalRowname}
\title{Make legal row/column names}
\usage{
makeLegalRowname(mets)
}
\arguments{
\item{x}{character vector of names to make legal}
}
\value{
The legal row/column names
}
\description{
make legal row/column names
}
\examples{
rown <- makeLegalRowname(mets);
}
