% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatAnalysis.R
\name{orderNCut}
\alias{orderNCut}
\title{Order and cut}
\usage{
orderNCut(x, threshold, N)
}
\arguments{
\item{x}{The data matrix or data frame}

\item{threshold}{threshold where measures be count as abnormal if its absolute value is above the or below the threshold}

\item{N}{maximum number of features to output}
}
\value{
The abnormal rate
}
\description{
Decreasingly order absolute values of measures and output top perturbed N features
}
\examples{
top30 <- getPertubedByColumn(data_mx, threshold=2,N=30);
}
