% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatAnalysis.R
\name{performTtestsAllClassesOneVsRest}
\alias{performTtestsAllClassesOneVsRest}
\title{get t-test result}
\usage{
performTtestsAllClassesOneVsRest(dataMatrix, classVector)
}
\arguments{
\item{dataMatrix}{The data matrix}

\item{classVector}{vector matching column names to categories}
}
\description{
get t-test result by group, one versus the rest
}
