% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatAnalysis.R
\name{performTtestsAllRows}
\alias{performTtestsAllRows}
\title{get t-test result}
\usage{
performTtestsAllRows(dataGroup1, dataGroup2)
}
\arguments{
\item{dataMatrix}{The data matrix}

\item{classVector}{vector matching colnames to catagories}
}
\value{
A data frame of mean differences and p-values
}
\description{
perform profile-wise t-test
}
\examples{
t.result = performTtestsAllClassesOneVsRest(dataMatrix=data_mx,classVector=c("treatment1","treatment2","treatment1","control"));;
}
