% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottings.R
\name{plotBoxPlot}
\alias{plotBoxPlot}
\title{plot box plots}
\usage{
plotBoxPlot(
  data_mx,
  sampleAttr,
  MOI,
  header4x,
  header4color = NULL,
  header4ID,
  palette = "Dark2"
)
}
\arguments{
\item{data_mx}{\itemize{
\item Normalized, imputed, z-scored data. Data matrix includes features as rows, samples as columns.
}}

\item{sampleAttr}{\itemize{
\item sample attributes; meta data
}}

\item{MOI}{\itemize{
\item molecule of interest. A subset of data_mx colnames.
}}

\item{header4x}{\itemize{
\item column name of sampelAttr used for x-axis categories.
}}

\item{header4color}{\itemize{
\item column name of sampelAttr used for color filling
}}

\item{header4ID}{-column name of sampelAttr indicating sample IDs.}

\item{palette}{color palette}
}
\description{
plot box plots
}
