% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottings.R
\name{plotCor}
\alias{plotCor}
\title{plot correlation between phenotypes or meta-level traits}
\usage{
plotCor(sampleAttr, type = c("pearson", "spearman"))
}
\arguments{
\item{sampleAttr}{metadata table with phenotype data listed by columns}

\item{type}{correlation method}
}
\description{
plot correlation between phenotypes or meta-level traits
}
