% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottings.R
\name{plotHeatmap}
\alias{plotHeatmap}
\title{#' plot blown-out modules in graph
#'
#' @param ig - igraph
#' @param S - a set of node names
#' @param header4x - column name of sampelAttr used for x-axis categories.
#' @param header4color - column name of sampelAttr used for color filling
#' @param header4ID -column name of sampelAttr indicating sample IDs.
#' @param palette color palette
#' @import visNetwork
#' @importFrom reshape2 melt
#' @example
#' @export
plotGraph=function(ig,S,size=NULL,membership,color=NULL,visNetworkOP=F,showLable=T){
if()}
\usage{
plotHeatmap(
  data_mx,
  classVector,
  classColorMatchList,
  colorRowSide = F,
  rowClassVector = NULL,
  heatColors = c("white", "steelblue"),
  colorby = c("value", "rank"),
  colorThresh = NULL,
  dendrogram = c("both"),
  Rowv = T,
  Colv = T,
  labRow = T,
  labCol = T
)
}
\description{

}
\details{
plot heatMap
}
