% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottings.R
\name{plotScatterPlot}
\alias{plotScatterPlot}
\title{plot scatter plots}
\usage{
plotScatterPlot(
  data_mx,
  sampleAttr,
  MOI,
  header4x,
  header4color = NULL,
  header4ID,
  palette = "Dark2"
)
}
\description{
plot scatter plots
}
