% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataPreprocess.R
\name{timeStamp}
\alias{timeStamp}
\title{timeStamp}
\usage{
timeStamp(date = F)
}
\arguments{
\item{date}{Boolean. Show date?}
}
\value{
time stamp
}
\description{
get current time stamp
}
\examples{
timeStamp();
}
