% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataPreprocess.R
\name{formatNumDataset}
\alias{formatNumDataset}
\title{Format and convert data set to numeric matrix}
\usage{
formatNumDataset(x, useRownames = F)
}
\arguments{
\item{x}{\itemize{
\item Data matrix
}}

\item{useRownames}{\itemize{
\item True row names need to be changed to feature names
}}
}
\value{
numeric data matrix
}
\description{
Format and convert data set to numeric matrix
}
