% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataPreprocess.R
\name{formatNumDataset}
\alias{formatNumDataset}
\title{Format and numerize dataset}
\usage{
formatNumDataset(x, useRownames = F)
}
\arguments{
\item{x}{\itemize{
\item Data matrix
}}

\item{useRownames}{\itemize{
\item True row names need to be changed to feature names
}}
}
\value{
numerated data matrix
}
\description{
Format and numerize dataset
}
