% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GraphModel.R
\name{getCTDmodule}
\alias{getCTDmodule}
\title{getCTDmodule}
\usage{
getCTDmodule(
  ig,
  profile,
  kmx = 30,
  useRanks = T,
  ranks = NULL,
  useS = F,
  S = NULL,
  p1 = 0.9,
  thresholdDiff = 0.01
)
}
\arguments{
\item{ig}{\itemize{
\item background network (condition-specific network).
}}

\item{profile}{\itemize{
\item named vector of sample omic profile named by corresponding feature name. If useS set as TRUE and S is provided, kmx can be NULL.
}}

\item{kmx}{\itemize{
\item number of top perturbed features to use as input node set. If useS set as TRUE and S is provided, kmx can be NULL.
}}

\item{useRanks}{\itemize{
\item Boolean. Whether to use user provided ranks.
}}

\item{ranks}{\itemize{
\item if useRanks is set as TRUE, provide calculated ranks here.
}}

\item{useS}{\itemize{
\item Boolean. If TRUE, provide S as input node set.
}}

\item{S}{\itemize{
\item vector of input node set.
}}

\item{p1}{\itemize{
\item parameter passed on to singleNode.getNodeRanksN. if useRanks is TRUE and ranks is provided then this value is ignored.
}}

\item{thresholdDiff}{\itemize{
\item parameter passed on to singleNode.getNodeRanksN. if useRanks is TRUE and ranks is provided then this value is ignored.
}}
}
\value{
a list of best compressed node set, p value.
}
\description{
get best compressed subset of input nodes and out put significance
}
