% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatAnalysis.R
\name{performTtestsAllClassesEachPair}
\alias{performTtestsAllClassesEachPair}
\title{get t-test result by group, pairwise}
\usage{
performTtestsAllClassesEachPair(dataMatrix, classVector)
}
\description{
get t-test result by group, pairwise
}
\section{perform t-test by group, in a pair wise manner}{
NA
}

