% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatAnalysis.R
\name{performTtestsAllClassesOneVsRest}
\alias{performTtestsAllClassesOneVsRest}
\title{get t-test result by group, one versus the rest}
\usage{
performTtestsAllClassesOneVsRest(dataMatrix, classVector)
}
\description{
get t-test result by group, one versus the rest
}
\section{perform t-test by group, one versus the rest of all groups}{
NA
}

